frontend/requirements.txt
streamlit==1.38.0
requests==2.32.3
pandas==2.2.3
sseclient-py==1.8.0
python-dotenv==1.0.1
